@@ -1,33 +1,34 @@
 {
     "name": "kadirov/api-starter-kit",
     "type": "project",
-    "license": "proprietary",
+    "license": "GPL-3.0-or-later",
     "require": {
         "php": "8.0.*",
         "ext-ctype": "*",
         "ext-iconv": "*",
-        "api-platform/core": "2.5.*",
-        "composer/package-versions-deprecated": "1.11.*",
-        "doctrine/doctrine-bundle": "2.2.*",
-        "doctrine/doctrine-migrations-bundle": "3.0.*",
+        "api-platform/core": "2.6.3",
+        "composer/package-versions-deprecated": "*",
+        "doctrine/doctrine-bundle": "2.3.*",
+        "doctrine/doctrine-migrations-bundle": "3.1.*",
         "doctrine/orm": "2.8.*",
         "lexik/jwt-authentication-bundle": "2.11.*",
         "nelmio/cors-bundle": "2.1.*",
-        "symfony/asset": "5.1.*",
-        "symfony/console": "5.1.*",
-        "symfony/dotenv": "5.1.*",
-        "symfony/expression-language": "5.1.*",
+        "symfony/asset": "5.2.*",
+        "symfony/console": "5.2.*",
+        "symfony/dotenv": "5.2.*",
+        "symfony/expression-language": "5.2.*",
         "symfony/flex": "1.12.*",
-        "symfony/framework-bundle": "5.1.*",
-        "symfony/http-client": "5.1.*",
-        "symfony/security-bundle": "5.1.*",
-        "symfony/twig-bundle": "5.1.*",
-        "symfony/validator": "5.1.*",
-        "symfony/yaml": "5.1.*"
+        "symfony/framework-bundle": "5.2.*",
+        "symfony/http-client": "5.2.*",
+        "symfony/security-bundle": "5.2.*",
+        "symfony/twig-bundle": "5.2.*",
+        "symfony/validator": "5.2.*",
+        "symfony/yaml": "5.2.*",
+        "symfony/property-info": "5.2.*"
     },
     "require-dev": {
         "roave/security-advisories": "dev-master",
-        "symfony/maker-bundle": "1.29.*"
+        "symfony/maker-bundle": "*"
     },
     "config": {
         "optimize-autoloader": true,
@@ -73,7 +74,7 @@
     "extra": {
         "symfony": {
             "allow-contrib": false,
-            "require": "5.1.*"
+            "require": "5.2.*"
         }
     }
 }
